requirements.txt
requests
python-dotenv
duckdb
dbt-core
dbt-duckdb
pandas
matplotlib
seaborn
notebook
